Create Coffee capsules
Some professors of DCX love to drink coffee and decided to buy a new special coffee machine for the department room. The coffee machine uses coffee capsules of many flavors to prepare the drink, and each capsule serves two cups of coffee

The capsules are sold in small boxes (10 units) or big boxes(16 units), and it was agreed that each professor would donate as many boxes as they want for the group. 

Write a program which receives the amount and the size of the boxes donated by each of the seven professors who share the coffee machine as input, and shows the total amount of coffee capsules donated, and how many cups each professor can drink.